requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
scikit-learn>=1.3.0
nltk>=3.8
plotly>=5.15.0
reportlab>=4.0.0
joblib>=1.3.0
numpy>=1.24.0
sentence-transformers>=2.7.0
spacy>=3.7.0
pdfplumber>=0.10.0
python-docx>=1.1.0
google-generativeai>=0.5.0
requests>=2.31.0
python-dotenv>=1.0.0
langchain-google-genai>=1.0.0
langchain-core>=0.2.0